refactor(auth): extract LoginLocal from bootstrap.go into login.go
